Refactored common part for ss-scale-* CLIs into slipstream.command.VerticalScaleCommanBase.
