+++ client/src/main/python/ss-scale-disk.py
@@ -19,53 +19,14 @@
 from __future__ import print_function
 
 import sys
-from slipstream.NodeDecorator import NodeDecorator
-from slipstream.SlipStreamHttpClient import DomExtractor
-
-from slipstream.command.CommandBase import CommandBase
-from slipstream.HttpClient import HttpClient
-import slipstream.util as util
+from slipstream.command.VerticalScaleCommandBase import VerticalScaleCommandBase
 from slipstream.wrappers.BaseWrapper import BaseWrapper
 
-etree = util.importETree()
 
-
-class MainProgram(CommandBase):
-    """A command-line program to request vertical scaling of node instances
-    in a deployment.
-    """
+class MainProgram(VerticalScaleCommandBase):
 
     def __init__(self, argv=None):
-        self.run_url = None
-        self.run_dom = None
-        self.node_name = None
-        self.instances_to_scale = []
-        self.rtp_scale_values = {}
-        self._scale_state = None
-        self._need_cloudservice_name = True
-
-        self.options = None
-        self.args = None
-
-        self.username = None
-        self.password = None
-        self.cookie = None
-        self.endpoint = None
         super(MainProgram, self).__init__(argv)
-
-    def parse(self):
-        usage = """usage: %prog [options] [--attach <GB> | --detach <device>] <run> <node-name> <ids> [<ids> ...]
-<run>        Run ID. Run should be mutable and in Ready state.
-<node-name>  Node name to scale the instances of.
-<ids>        IDs of the node instances to scale."""
-
-        self.parser.usage = usage
-        self.add_authentication_options()
-        self.addEndpointOption()
-        self.add_scale_options()
-
-        self.options, self.args = self.parser.parse_args()
-        self._check_args()
 
     def add_scale_options(self):
         """INTERFACE
@@ -75,23 +36,7 @@ class MainProgram(CommandBase):
         self.parser.add_option('--detach', dest='device', default=None,
                                help='Device name to detach. (Example: /dev/vdc)', metavar='DEVICE')
 
-    def _check_args(self):
-        if len(self.args) < 3:
-            self.usageExitTooFewArguments()
-        run_id = self.args[0]
-        self.run_url = self.options.endpoint + util.RUN_RESOURCE_PATH + '/' + run_id
-        self.node_name = self.args[1]
-
-        self._validate_and_set_scale_options()
-
-        try:
-            self.instances_to_scale = map(int, self.args[2:])
-        except ValueError:
-            self.usageExit("Invalid ids, they must be integers")
-
     def _validate_and_set_scale_options(self):
-        """INTERFACE
-        """
         if not any([self.options.attach_gb, self.options.device]):
             self.usageExit("Either --attach or --detach should be provided")
 
@@ -106,80 +51,8 @@ class MainProgram(CommandBase):
             self.rtp_scale_values['disk.detach.device'] = self.options.device
             self._scale_state = BaseWrapper.SCALE_STATE_DISK_DETACHING
 
-    def _get_scale_state(self):
-        """INTERFACE
-        """
-        return self._scale_state
-
     def doWork(self):
-
-        client = HttpClient(self.options.username, self.options.password)
-        client.verboseLevel = self.verboseLevel
-
-        self._retrieve_and_set_run(client)
-
-        self._check_allowed_to_scale(client)
-
-        self._set_scale_rtps(client)
-
-        self._set_scale_state(client)
-        self._set_run_provisioning(client)
-
-    def _set_scale_rtps(self, client):
-        self.log("Requesting to scale node instances: %s" % self.instances_to_scale)
-
-        cloudservice_name = self._get_cloudservice_name()
-        cloudservice_name_with_separator = cloudservice_name and (cloudservice_name + '.') or ''
-
-        node_url = self.run_url + "/" + self.node_name
-        for _id in self.instances_to_scale:
-            url = node_url + NodeDecorator.NODE_MULTIPLICITY_SEPARATOR + \
-                str(_id) + NodeDecorator.NODE_PROPERTY_SEPARATOR + cloudservice_name_with_separator
-            for scale_key, value in self.rtp_scale_values.items():
-                print(url + scale_key)
-                client.put(url + scale_key, value)
-
-    def _get_cloudservice_name(self):
-        if self._need_cloudservice_name:
-            run_params = DomExtractor.extract_run_parameters_from_run(self.run_dom)
-            return run_params[self.node_name + NodeDecorator.NODE_PROPERTY_SEPARATOR +
-                NodeDecorator.CLOUDSERVICE_KEY]
-        else:
-            return ''
-
-    def _set_scale_state(self, client):
-        node_url = self.run_url + "/" + self.node_name
-        for _id in self.instances_to_scale:
-            url = node_url + NodeDecorator.NODE_MULTIPLICITY_SEPARATOR + \
-                str(_id) + NodeDecorator.NODE_PROPERTY_SEPARATOR + 'scale.state'
-            client.put(url, self._get_scale_state())
-
-    def _set_run_provisioning(self, client):
-        client.put(self.run_url + '/ss:state', 'Provisioning')
-
-    def _check_allowed_to_scale(self, client):
-        err_msg = "ERROR: Run should be mutable and in Ready state."
-
-        ss_state = self._get_ss_state(client)
-        if 'Ready' != ss_state:
-            self.usageExit(err_msg + " Run is in %s state." % ss_state)
-
-        if not self._is_run_mutable():
-            self.usageExit(err_msg + " Run is not mutable.")
-
-    def _get_ss_state(self, client):
-        ss_state_url = self.run_url + "/" + NodeDecorator.globalNamespacePrefix + 'state'
-        _, ss_state = client.get(ss_state_url)
-        return ss_state
-
-    def _retrieve_and_set_run(self, client):
-        _, run_xml = client.get(self.run_url, 'application/xml')
-        self.run_dom = etree.fromstring(run_xml)
-
-    def _is_run_mutable(self):
-        mutable = DomExtractor.extract_mutable_from_run(self.run_dom)
-        return util.str2bool(mutable)
-
+        super(MainProgram, self).doWork()
 
 if __name__ == "__main__":
     try:
